requirements.txt
Flask==0.10.1
httplib2==0.9.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
oauth2==1.5.211
phonenumbers==7.0.9
python-instagram==1.3.2
python-yelp-v2==0.5.5
simplejson==3.8.0
six==1.9.0
Werkzeug==0.10.4
wheel==0.24.0
